feat: v0.7.0 — RTL, full locales, PHPStan-clean, docs
Milestone 4 (final parity & polish).

RTL:
- master.blade.php loads the prebuilt adminlte.rtl.min.css when layout_rtl
  is enabled; installer publishes it from node_modules to public/vendor.

Locales:
- 6 new fully-translated locales (fr, it, pt_BR, ru, zh, ja) and de/es
  brought to parity — all 9 locales now cover the same 129 keys.
- Deduplicated redundant keys in the canonical en file.

Quality:
- PHPStan level 8 is now fully clean (0 errors). Added accurate generic
  array typehints across menu, filters, plugins, and components; made
  json_encode-based config methods return string via `?: '{}'`; fixed the
  container offset access in AdminLte; scoped-ignored the package
  view-string false-positive (views are namespaced at runtime).

Status command checks the RTL stylesheet, the four plugin vendor files,
and scaffolded sections. install --only=lang publishes language files.

Docs: README gains a Scaffolding + Authentication section; CHANGELOG
updated to reflect the real DB-backed scaffolding and the asset-pipeline
fixes.

Verified end-to-end: built the demo app and captured a headless-Chrome
screenshot — the dashboard renders the full AdminLTE design (live
ApexCharts chart, info/small boxes, kanban, calendar, progress bars,
accordion), confirming the earlier "wireframe" issue is resolved. Pint
clean, PHPStan 0 errors, 26 PHPUnit tests pass.

+++ src/View/Components/Widget/ProgressGroup.php
@@ -17,7 +17,7 @@ class ProgressGroup extends Component
 
     public function percentage(): int
     {
-        return round(($this->value / ($this->max ?? 100)) * 100);
+        return (int) round(($this->value / ($this->max ?? 100)) * 100);
     }
 
     public function render(): View
